Separate analyzers and code fixes.
