+++ PosInformatique.Moq.Analyzers.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 18
-VisualStudioVersion = 18.4.11626.88 stable
+VisualStudioVersion = 18.4.11626.88
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Moq.Analyzers.Tests", "tests\Moq.Analyzers.Tests\Moq.Analyzers.Tests.csproj", "{DEE86A7E-8338-4C3D-822B-E8FB976D9905}"
 EndProject
@@ -81,6 +81,17 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{02EA681E-C7D
 		src\.editorconfig = src\.editorconfig
 	EndProjectSection
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Moq.Analyzers.CodeFixes", "src\Moq.Analyzers.CodeFixes\Moq.Analyzers.CodeFixes.csproj", "{310BC90C-AC67-4DAB-86A9-76E7DE8B1E40}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{A0CD9438-DC28-48A0-8CF4-B2A809546FF6}"
+	ProjectSection(SolutionItems) = preProject
+		src\Directory.Build.props = src\Directory.Build.props
+	EndProjectSection
+EndProject
+Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "Moq.Analyzers.Shared", "src\Moq.Analyzers.Shared\Moq.Analyzers.Shared.shproj", "{8C28B748-EF68-4685-A901-49BE72FCAC4C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Moq.Analyzers.CodeFixes.Tests", "tests\Moq.Analyzers.CodeFixes.Tests\Moq.Analyzers.CodeFixes.Tests.csproj", "{44AE11BA-BD29-4503-9479-8BD77E99E6CF}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -99,6 +110,14 @@ Global
 		{07F970A1-1477-4D4C-B233-C9B4DA6E3AD6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{07F970A1-1477-4D4C-B233-C9B4DA6E3AD6}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{07F970A1-1477-4D4C-B233-C9B4DA6E3AD6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{310BC90C-AC67-4DAB-86A9-76E7DE8B1E40}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{310BC90C-AC67-4DAB-86A9-76E7DE8B1E40}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{310BC90C-AC67-4DAB-86A9-76E7DE8B1E40}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{310BC90C-AC67-4DAB-86A9-76E7DE8B1E40}.Release|Any CPU.Build.0 = Release|Any CPU
+		{44AE11BA-BD29-4503-9479-8BD77E99E6CF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{44AE11BA-BD29-4503-9479-8BD77E99E6CF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{44AE11BA-BD29-4503-9479-8BD77E99E6CF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{44AE11BA-BD29-4503-9479-8BD77E99E6CF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -110,8 +129,14 @@ Global
 		{D9C84D36-7F9C-4EFB-BE6F-9F7A05FE957D} = {3C20D95F-AB5F-44EC-8FB6-CB9827B7FD63}
 		{FA7258E1-65F0-4C40-A122-D76A1F0C79A0} = {1D59B801-B4D3-44FC-A2BE-F2F53AC54061}
 		{0EACA365-AFB2-4200-91FD-A20DC9617799} = {FA7258E1-65F0-4C40-A122-D76A1F0C79A0}
+		{A0CD9438-DC28-48A0-8CF4-B2A809546FF6} = {1D59B801-B4D3-44FC-A2BE-F2F53AC54061}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {3307E7F7-9CD7-4C12-B34A-943F5A8B62A4}
+	EndGlobalSection
+	GlobalSection(SharedMSBuildProjectFiles) = preSolution
+		src\Moq.Analyzers.Shared\Moq.Analyzers.Shared.projitems*{1962bef9-e6df-4485-a113-e255c84177d4}*SharedItemsImports = 5
+		src\Moq.Analyzers.Shared\Moq.Analyzers.Shared.projitems*{310bc90c-ac67-4dab-86a9-76e7de8b1e40}*SharedItemsImports = 5
+		src\Moq.Analyzers.Shared\Moq.Analyzers.Shared.projitems*{8c28b748-ef68-4685-a901-49be72fcac4c}*SharedItemsImports = 13
 	EndGlobalSection
 EndGlobal
