+++ src/Moq.Analyzers.Shared/ChainInvocationArgument.cs
