+++ src/Moq.Analyzers.Shared/ChainMember.cs
