+++ src/Moq.Analyzers.Shared/ChainMembersInvocation.cs
