+++ src/Moq.Analyzers.Shared/Moq.Analyzers.Shared.projitems
@@ -1,0 +1,22 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <MSBuildAllProjects Condition="'$(MSBuildVersion)' == '' Or '$(MSBuildVersion)' &lt; '16.0'">$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
+    <HasSharedItems>true</HasSharedItems>
+    <SharedGUID>8c28b748-ef68-4685-a901-49be72fcac4c</SharedGUID>
+  </PropertyGroup>
+  <PropertyGroup Label="Configuration">
+    <Import_RootNamespace>Moq.Analyzers.Shared</Import_RootNamespace>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="$(MSBuildThisFileDirectory)AnalyzerRules.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ChainInvocationArgument.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ChainMember.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)ChainMembersInvocation.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Diagnostics\NotNullWhenAttribute.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)MoqExpressionAnalyzer.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)MoqSymbols.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)RaiseMethodCall.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)SyntaxNodeHelper.cs" />
+  </ItemGroup>
+</Project>
