+++ src/Moq.Analyzers.Shared/MoqSymbols.cs
