+++ src/Moq.Analyzers.Shared/SyntaxNodeHelper.cs
