+++ src/Moq.Analyzers/Analyzers/SetupSetAnalyzer.cs
@@ -15,8 +15,8 @@ namespace PosInformatique.Moq.Analyzers
     [DiagnosticAnalyzer(LanguageNames.CSharp)]
     public class SetupSetAnalyzer : DiagnosticAnalyzer
     {
-        internal static readonly DiagnosticDescriptor UseSetupSetWithGenericArgumentRule = new DiagnosticDescriptor(
-            "PosInfoMoq1005",
+        private static readonly DiagnosticDescriptor UseSetupSetWithGenericArgumentRule = new DiagnosticDescriptor(
+            AnalyzerRules.UseSetupSetWithGenericArgumentRuleId,
             "Defines the generic argument of the SetupSet() method with the type of the mocked property",
             "Defines the generic argument of the SetupSet() method with the type of the mocked property",
             "Design",
