+++ tests/Moq.Analyzers.CodeFixes.Tests/CodeFixes/AddVerifyAllCodeFixProviderTest.cs
@@ -4,7 +4,7 @@
 // </copyright>
 //-----------------------------------------------------------------------
 
-namespace PosInformatique.Moq.Analyzers.Tests
+namespace PosInformatique.Moq.Analyzers.CodeFixes.Tests
 {
     using Verifier = MoqCSharpCodeFixVerifier<VerifyAllShouldBeCalledAnalyzer, AddVerifyAllCodeFixProvider>;
