+++ tests/Moq.Analyzers.Sandbox/Moq.Analyzers.Sandbox.csproj
@@ -6,17 +6,10 @@
 
   <ItemGroup>
     <PackageReference Include="Moq" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" />
-    <PackageReference Include="xunit" />
-    <PackageReference Include="xunit.runner.visualstudio">
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-      <PrivateAssets>all</PrivateAssets>
-    </PackageReference>
-    
-    <PackageReference Remove="StyleCop.Analyzers" />
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\src\Moq.Analyzers.CodeFixes\Moq.Analyzers.CodeFixes.csproj" />
     <ProjectReference Include="..\..\src\Moq.Analyzers\Moq.Analyzers.csproj" PrivateAssets="all" ReferenceOutputAssembly="false" OutputItemType="Analyzer" Condition="'$(Configuration)' == 'Debug'" />
   </ItemGroup>
